+++ src/components/layout/SideBar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import Image from './ui/Image';
-import Button from './ui/Button';
+import Image from '../ui/Image';
+import Button from '../ui/Button';
 import Profile from '@/assets/profile.svg?url';
 
 const SideOverlay = styled.div`
